Refactored edit translational modal
