+++ src/components/translation-builder/translation-builder.module.scss
@@ -28,7 +28,7 @@
   grid-template-columns: 1fr 1fr auto;
   margin-bottom: layout.$spacing-05;
   padding: layout.$spacing-03 0;
-  border-bottom: 1px solid #e0e0e0;
+  border-bottom: 1px solid $ui-03;
   background-color: #fff;
 
   .translationKey {
