+++ src/routes.json
@@ -70,6 +70,10 @@
     {
       "name": "delete-form-modal",
       "component": "deleteFormModal"
+    },
+    {
+      "name": "edit-translation-modal",
+      "component": "editTranslationModal"
     }
   ]
 }
