+++ translations/en.json
@@ -8,6 +8,7 @@
   "addQuestion": "Add Question",
   "addReference": "Add Reference",
   "addSection": "Add Section",
+  "all": "All",
   "answerConceptValidation": "One or more selected answer concepts do not exist in the system. ",
   "auditDetails": "Audit Details",
   "autogeneratedUuid": "UUID (auto-generated)",
@@ -54,12 +55,16 @@
   "description": "Description",
   "descriptionPlaceholderText": "e.g. A form used to collect encounter data for clients in the Express Care program.",
   "downloadSchema": "Download schema",
+  "downloadTranslation": "Download translation",
   "dragToReorder": "Drag to reorder",
   "duplicateQuestion": "Duplicate question",
   "editButton": "Edit {{elementType}}",
+  "editingFor": "Editing translation for:",
   "editQuestion": "Edit question",
   "editSchema": "Edit schema",
   "editSection": "Edit section",
+  "editString": "Edit string",
+  "editTranslationTitle": "Edit Translation",
   "encounterType": "Encounter Type",
   "enterPageTitle": "Enter a title for your new page",
   "enterSectionTitle": "Enter a section title",
@@ -131,6 +136,7 @@
   "loading": "Loading",
   "loadingPreview": "Loading preview...",
   "loadingSchema": "Loading schema",
+  "loadingTranslations": "Loading translations...",
   "manageForms": "Manage Forms",
   "mappings": "Mappings",
   "maximizeEditor": "Maximize editor",
@@ -143,6 +149,7 @@
   "noMatchingConcepts": "No concepts were found that match",
   "noMatchingFormsToDisplay": "No matching forms to display",
   "noSchemaLoaded": "No schema loaded",
+  "noTranslations": "No translatable strings found.",
   "notSearchable": "Not Searchable",
   "noValidationErrorsFound": "No validation errors found",
   "off": "Off",
@@ -225,12 +232,16 @@
   "select": "Select",
   "selectAnswersToDisplay": "Select answers to display",
   "selectForm": "Select Form",
+  "selectLanguage": "Select language",
   "selectProgram": "Select a program",
   "selectQuestions": "Select Questions",
   "source": "Source",
   "startBuilding": "Start building",
   "success": "Success!",
   "timerOnly": "Timer only",
+  "translated": "Translated",
+  "translationBuilder": "Translation Builder",
+  "translationValue": "Translated Value",
   "tryAgain": "Try again",
   "typeRequired": "Type is required",
   "undo": "Undo",
@@ -240,6 +251,7 @@
   "unpublishConfirmation": "Are you sure you want to unpublish this form?",
   "unpublishExplainerText": "Unpublishing a form means you can no longer access it from your frontend. Unpublishing forms does not delete their associated schemas, it only affects whether or not you can access them in your frontend.",
   "unpublishForm": "Unpublish Form",
+  "untranslated": "Untranslated",
   "update": "Update",
   "updateExistingForm": "Update existing version",
   "validateAndPublishForm": "Validate and publish form",
